requirements.txt
aiogram==3.3.0
pandas
python-dotenv
google-generativeai
fuzzywuzzy[speedup]
